Removed useeffect from notifications logic implementation
